e_commerce_crawler/requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.2
pyyaml>=6.0.1
aiosignal>=1.3.1
charset-normalizer>=3.1.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
urllib3>=2.0.0
requests>=2.31.0